backend/requirements.txt
Flask==3.0.2
flask-cors==4.0.0
gunicorn==21.2.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
pillow==10.2.0
scikit-image==0.22.0
joblib==1.3.2
tensorflow==2.17.0
setuptools>=68.0.0
wheel>=0.40.0
